requirements.txt
Django==4.2.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-filter==23.3
schedule==1.2.0
python-decouple==3.8 